+++ config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  //number of pages to write at a time, to the queue
+  "batch_size": 1000,
+  //the default name of the collection to write to
+  "collection": "wikipedia",
+  //update interval
+  "logInterval": 1000,
+}