+++ src/worker/_polyfill.js
@@ -1,0 +1,14 @@
+
+//clearer than indexOf - string.includes is node>=6.3
+if (!String.prototype.includes) {
+  String.prototype.includes = function(search, start) {
+    if (typeof start !== 'number') {
+      start = 0;
+    }
+    if (start + search.length > this.length) {
+      return false;
+    } else {
+      return this.indexOf(search, start) !== -1;
+    }
+  };
+}