+++ changelog.md
@@ -10,3 +10,9 @@
 ### v2.4.0
 * add a 3s 'break' to avoid build-up of mongo inserts
 * add new --verbose and --skip_first options
+
+## v3
+* MASSIVE SPEEDUP! full re-write by @devrim 🙏 to fix #59
+* rename from `wikipedia-to-mongo` to `dumpster-dive`
+* use wtf_wikipedia v3 (a big re-factor too!)
+* use `line-by-line`, and `worker-nodes` to run parsing in parallel